requirements.txt
django~=4.2
pillow~=9.5.0
base58~=2.1.1
